ambiente-final/projeto/backend_python/requirements.txt
psycopg2-binary
flask
requests
PyJWT
cryptography
bcrypt
flask_jwt_extended
gunicorn